Utils: Odometry & sequances
+ Odometry read calibration files with function in 'utils' folder.
+ Function for sequances generation are in 'utils/kitti.py'.
